Possible Statcap fix for issue https://github.com/Sphereserver/Source/issues/121
The issue happens in the method Stat_Decrease() (CCharStat.cpp), this method decrease a stat (if set with the stat arrow down) when another stat
is increasing and the player is at Statcap.

In the following code:
`
 int iStatSumAvg = Stat_GetLimit(STAT_QTY);
 int iStatSum = Stat_GetSum() + 1; // +1 here assuming we are going to have +1 stat at some point thus we are calling this function
 if (iStatSum < iStatSumAvg)  //No need to lower any stat.
  return true;
 ...code for decreasing...
`
If the sum of the player stats (Stat_GetSum) is a Statcap - 1, the sum of the player stats will be considered at the same value
of the Statcap ( int iStatSum = Stat_GetSum() + 1), so the check in the line below will not pass and the Stat will increase only if one of the other stats
is set for decreasing (stat arrow down) even if the player actual stat sum is at Stat Cap - 1.
So i changed the check:
if (iStatSum < iStatSumAvg) to if (iStatSum <= iStatSumAvg)

That means i  also needed to change the following condition in the method Skill_Experience() (CCharSkill.cpp):
`
if (iStatSum >= iStatCap) // stat cap already reached
 break
`
to
`
 if (iStatSum > iStatCap)  // stat cap already reached
 break
`
The first condition will skip the stat gain when the player is above or equal the stat cap, but it will also prevent the player to decrease his stats
by using the stat arrows buttons when the player have reached the Statcap.