Added LASTEVENTWALK, CONTAINERMAXITEMS, fixed switching weapon glitch and changed default compiler architecture to 64 bits, instead of 32 bits.
Added: LASTEVENTWALK read-only client property, to retrieve the time in ticks (SERV.TIME) of the last allowed walk request from the client.
Added: CONTAINERMAXITEMS ini setting. It holds the maximum number of items inside a container, which is by default (and can be at most) 255. Higher values are unsupported by the Classic Client and the older 2D client versions. The Enhanced Client supports a maximum af 125 items per container, so it's suggested to update this setting if you are going to use that client.
Fixed: Switching weapon while attacking didn't abort the previous swing (Issue #108).