Fixes regarding NPC combat AI and garbage collection.
- Fixed: Items in trade window being lost if a world save occurred during the trade process and the server was restarted, loading that save.
- Fixed: NPCs sometimes lost the focus on their enemy during combat.

NPCs not always attacked the targeted char when ordered by a GM to do so, especially if it was of the same type. This doesn't happen when a player does so because the attacker has the flag STATF_PET. This issue isn't 100% solved yet, but this commit makes it happen less often.