Reverted TICKS_PER_SEC to 10, fixed PeriodicTick exception, improvements to TimerF.
Since ticks now are relevant only for periodic ticking and not for worldobject timers, we can afford to revert TICKS_PER_SEC to 10. WorldObjects timers are now being checked and eventually fired/ticked with a millisecond (or more) precision, they aren't related anymore to the world tick, which happened every 0,1s or 0,25s. Also, the TIMERT keyword is kinda useless now, so maybe it'll be removed in the near future.

Fixed exception occurring in char's periodic ticking code when a PeriodicTick(ing) char wasn't removed via CChar::Delete() method.
Added several checks to TimerF code, which may help with the Issue #131. Also fixed high-value UIDs saved in TimerFNumbers line (spheredata.scp) not being correctly loaded (atoi loses the last bit if we want an unsigned number).
Added informations to the debug messages shown in case an exception occurs when loading a script section.