Fix for Issue #248 regarding the AMMOANIM, AMMOANIMHUE, AMMOANIMRENDER, AMMOSOUNDHIT, AMMOSOUNDMISS and AMMOCONT properties.
I also made a fix for  AMMOCONT  but while it seems to work i am not sure if it's the right way, so i have left the old code commented.
The problem i had with AMMOCONT was because it can stores both a numeric value (uid or numeric id) or a string (the string id, ex: i_bag).

Also i have noticed that the Weapon_GetSoundHit/Weapon_GetSoundMiss methods are  called only when used against an archery butte.