Fixed MORE, added REGION_FLAG_NOMINING and REGION_FLAG_WALK_NOBLOCKHEIGHT. - Added: REGION_FLAG_NOMINING (020000) to prevent mining in a region. - Added: REGION_FLAG_WALK_NOBLOCKHEIGHT (040000) to ignore char height for walkchecks inside the region (see MT_NOBLOCKHEIGHT). Use with caution! - Fixed: MORE value wasn't saved correctly if the value was a defname. - Fixed: When retrieving MORE value, the numeric value was always returned even if it should have been a defname. Retrieving MORE1 didn't have this problem.
Added FALLTHROUGH macro to explicitly state our intentions when there's no break at the end of a case clause inside a switch.
Fixed infinite loop when a char couldn't be spawned to a given P because it would be stuck there.
Changed PLAYER_HEIGHT macro from 15 to 16, to match the default value in scripts.