Improved stack trace precision and fixed corpses appearing to change DIR after UPDATE (on SA+ clients).
- Changed: Improved a lot the precision of the stack trace, now it should point exactly (in most cases) to the function where the exception was generated.
- Fixed: UPDATE caused Stygian Abyss or later clients to show corpses changing their laying direction (even if it actually was a visual artifact, direction wasn't changing server-side) (Issue #274).

Avoided exception when dynamically setting an item to type t_multi_custom (you can set it only in the itemdef).
Renamed CException to CWinException, since it's used only by _set_se_translator on Windows.
In CChar::OnAttackedBy, reduced the calls to Memory_AddObjTypes from 2 to 1, so that less noto updates are sent to the client.