Updated custom multi AddStairs/RemoveItem (#330)
* Updated custom multi AddStairs/RemoveItem

Modified the loop to search through iStairID in AddStairs (though using a short, no idea if that's the correct naming convention), now it is similar to what you can see in RemoveStairs. The last argument seems to be related to AddMulti but couldn't find a reason to keep it.

RemoveItem was modified to not remove the entire floor no matter what, client doesn't seem to like that, the server might remove the floor but you can commit and see the whole invalid items placed on it or whatever was removed from view by the server. I removed that bunch of code, maybe it would be the best to put it between some #ifdef or leave it commented.

Anyway, these are the differences I found between the 56d and X versions.
Tested on latest client (7.0.69?). Everything seems to work fine.

* Commented old code instead of just removing it.