Fixed: Dismounting from a NPC cleared the actarg1 value of the rider instead of the NPC mount. (Issue #527)    This means that when unmounting and the character is using a skill that makes use of the Actarg1 value, such  as casting a spell, the skill will continue its execution normally. Fixed: Invalid spell ID didn't abort the skill execution when the magic skill succeeds (this is related to the issue above), causing a possible skill gain.
When a NPC mount is mounted, this NPC will holds the UID of the mount item (stored in the rider layer) in its Actarg1 value (as we can see in the Horse_GetMountItem method),
but when the NPC mount is dismounted the Actarg1 of the rider is cleared instead of the NPC mount.

This caused a problem when the rider was performing a skill that make uses of actarg1 (such as a casting a spell) and quickly dismounting:
In the case of a magic skill, actarg1 holds the spell ID, and by clearing this value when dismounting, it caused the spell to halt.
Combined with problem above, a wrong return value when the spell ID is invalid, in the Skill_Magery method, didn't abort the skill success execution, allowing a player to increase the
magic skill score without the spell consuming mana / reagents.