-Fixed: NPCs stop attacking after being poisoned NPC  (Issue #487). The bug happens when a NPC get damaged while the player is not reachable and it was caused by a missing assignement in the check in method CCharFight::OnTakeDamage `  if (bAttackerExists == false)   {    ....    attacker.ignore = false; //Attacking.ignore was not set causing it to be set at 246    m_lastAttackers.emplace_back(std::move(attacker));   } `
When the player goes away, the player is properly removed from the NPC's attacker list,  but if in the meantime the NPC suffers damage from the player by a command/item that
calls directly OnTakeDamage the NPC will add back the player but without setting the attacker.ignore property properly (It was set to 246 in my case).

-Fixed: Fixweight not working properly (Issue #502).

The problem was caused by the  method call:
`
 m_totalweight += pCont->GetWeight();
...
int CItemContainer::GetWeight(word amount) const
{ // true weight == container item + contents.
 return( CItem::GetWeight(amount) + CContainer::GetTotalWeight());
}
`
CContainer::GetTotalWeight would return m_totalweight of the current container but at the top of the Fixweight method we set m_totalweight value to 0 with the result
that Fixweight returned only the weight of the single item containers.

Minor problem i found:
When a player login, it seems that the Bank box weight it's added to is own weight, so a naked player should only weights 3 stones (backpack) but instead it weights 6 stones (3 backpack + 3 bank box)
I can't manage to track it.