Fix for Issue #569: NPC with stone flag keeps switching between war / peace mode.
There was no  "stone flag" check in the OnTick() method in CCharAct.cpp this allowed the NPC to  continue is normal behaviour and eventually reaching the Fight_CanHit method in CCharFight.cpp where it always returned  the invalid swing type causing the Fight_Clear method to be called.
This "loop" would continue until the player left the NPC line of sight.