Fix horses killed while ridden
Horses are set to z -127 //max 128
This way prevent any damage (i think) and let work timers and get scripted damage.
Setdisconnect() would prevent timers to work if it is not already prevented.
Stables works with disconnect method.
Non-tested yet. Could be done using .Stone  on the horse as well. so, maybe it is better option, but this way we only mod one lane