Fix: Deleting a player while it's in a guild cause the server to crash. (Issue #971).
When a player is deleted the ClearPlayer method is called and this set the m_pPlayer property to null. When a deleted player is inside a guild, the Guild_Resign method will fail because this method relies upon the m_pPlayer to succeed (and the destructor is called after ClearPlayer).

So i just moved the Guild_Resign method in the ClearPlayer before the m_pPlayer property is set to null.