Added mutex lock retry macros
Added: RETRY_LOCK_FOR_TASK*, used to retry mutex lock in the (rare) case
it fails. This has to be used only for critical mutexes.
Added: STDERR_LOG to log critical messages to stderr if standard CLog
methods are failing or can't be used.
Added: NOTHROW macro to specify functions that might throw an exception,
as opposite to nothrow.
Adapted some CLog and threading methods to actually be noexcept.
Reordered common.h.