Enhanced Objects Ticking System.
Added double checks to ensure that only valid objects remain in the
ticking lists, and that elapsed timeouts are actually removed from the
lists. This also should fix some NPCs not going to sleep while in a
sleeping sector.
Also, this new structure should be faster, since now it's not needed to
loop through the ticking lists elements to know if the given object is
already in some list.