Added new CombatFlag ini setting COMBAT_ATTACK_NOAGGREIVED (#1342)
* Added new CombatFlag ini setting COMBAT_ATTACK_NOAGGREIVED

In the old behavior (55i/55r/56b/56c) we were not guilty when we hit a murderer. Now when we attack a criminal or murderer we are guilty towards that person. An ini setting option has been added for this behavior. Although this is a normal behavior in OSI, it can be perceived as a problem since not every server designs games like OSI.

---------

Co-authored-by: cbnolok <cbnolok@gmail.com>