Review and merge PR #1319: OF_PetBehaviorOwnerNeutral, OF_NPCMovementOldStyle, TAG.OVERRIDE.MOVESTYLE.
- Added: New optionflag OF_PetBehaviorOwnerNeutral.
    If this setting is enabled, pets you own will appear to you with the
original noto (pre-taming, old behavior). If this setting is not
enabled, the assets you own will always appear natural to you. This
setting only applies when the owning character is looking at the asset
they own. It does not apply when someone else is looking at the asset
you own.
- Added: OF_NPCMovementOldStyle, TAG.OVERRIDE.MOVESTYLE.
    Support for the old 0.56b NPC movement formulas with:
OF_NPCMovementOldStyle ini setting and TAG.OVERRIDE.MOVESTYLE (to
override the behavior for a single character).
    With that formula, among the other things, char MOVERATE property
and TAG.OVERRIDE.MOVERATE will return to be the percent modifier of the
speed of the NPC - the higher it is, the slower will the NPC be. Default
is 100.